Add author_reply util
author_reply is a handler for sending a direct message to the author of the context message. If the user's DMs are closed, it handles the error appropriately by notifying the user in the server channel that their DM is closed. Supports the all options of discord abc.Messageable.send function. If sent successfully, returns the sent Message object