Some cogs (#38)
* Add commands to load, unload, and reload commands

Squash: Add find_command

* Change to easier way of importing util module functions

* Add delay, parseTime, and isNaN utilities

* Add 8ball cog

* Add choose cog

* Add restart command

Squash: Fix typo

* Add purge cog

* Add ping cog

* Update ping with client endpoint latency

* Add build_info cog

* Add help cog

* Add author_reply util

author_reply is a handler for sending a direct message to the author of the context message. If the user's DMs are closed, it handles the error appropriately by notifying the user in the server channel that their DM is closed. Supports the all options of discord abc.Messageable.send function. If sent successfully, returns the sent Message object

* Change help to be dynamically created based on command objects

* Change desc to description in command decorator

* Change uptime to be based on connection time instead of init time

Start time (bot.readyAt) is now updated whenever the on_ready event occurs