Change from requirements to Pipfile dependency tracking
Pipfile uses pipenv, which listed in the dev dependencies which is a thing it allows you to track. This is closer to the standards of other package managers.
The other thing is the lockfile (Pipfile.lock), which is a specific list of all of the packages with exact versions used in a commit, which allows determinism
in replicating an environment