Cog master (#42)
* Uptime.py (#24)

Moving to repo branch

* Reorganize cog implementation

* Cog implementation (#20)

Add cog loader

* Fix embed shrinkage

* Dynamically output time using only significant digits

* Make units correctly singular/plural when they should be

* Add standard python gitignore to gitignore

* Extract utility functions to utility modules

* Change error handler to ignore command not found

* Comment out general error handler

* music.py

* Delete music.py

* Add StringBuilder class for efficient string concatenation

* Move import statement

* Remove redundant newline

* Update syntax for commands in bot help

* music.py

Must install ffmpeg for Play command

* music.py Base

Too complex for me so we're building off of this for base

* Removal of music from bot.py

Moved to use of cog

* Profile kick ban cog (#33)

* Move profile, kick, and ban to cogs

* Round timestamps in profile to nearest second

* Change uptime.py to use StringBuilder

* Change prefix back to !

* Add special case to kick & ban for no input

* Update music.py

* Shinfo cog (#37)

* Add sh_info cog

* Revert spacing changes

* Some cogs (#38)

* Add commands to load, unload, and reload commands

Squash: Add find_command

* Change to easier way of importing util module functions

* Add delay, parseTime, and isNaN utilities

* Add 8ball cog

* Add choose cog

* Add restart command

Squash: Fix typo

* Add purge cog

* Add ping cog

* Update ping with client endpoint latency

* Add build_info cog

* Add help cog

* Add author_reply util

author_reply is a handler for sending a direct message to the author of the context message. If the user's DMs are closed, it handles the error appropriately by notifying the user in the server channel that their DM is closed. Supports the all options of discord abc.Messageable.send function. If sent successfully, returns the sent Message object

* Change help to be dynamically created based on command objects

* Change desc to description in command decorator

* Change uptime to be based on connection time instead of init time

Start time (bot.readyAt) is now updated whenever the on_ready event occurs

* Remove bot instantiation from music.py and move to commands folder

* Remove youtube_dl from bot.py imports

* Change from requirements to Pipfile dependency tracking

Pipfile uses pipenv, which listed in the dev dependencies which is a thing it allows you to track. This is closer to the standards of other package managers.
The other thing is the lockfile (Pipfile.lock), which is a specific list of all of the packages with exact versions used in a commit, which allows determinism
in replicating an environment

* Add Napkins name to README.md

Co-authored-by: NapkinsXYZ <65756491+NapkinsXYZ@users.noreply.github.com>